Finished home top items

+++ src/sql.ts
@@ -1,5 +1,5 @@
 import {Database, sqlite3} from "sqlite3";
-import {CameraFeed, Setting} from "./include/interfaces";
+import {CameraFeed, Recording, Setting} from "./include/interfaces";
 
 const {waitForFolderToExist} = require("./fileUtils");
 const sqlite3 = require('sqlite3').verbose();
@@ -7,6 +7,7 @@ let db: Database;
 
 const cameraFeedTableName = 'camera_feeds';
 const settingsTableName = 'settings';
+const recordingsTableName = 'recordings';
 
 export const createDbTables = (): void => {
     waitForFolderToExist('data').then(() => {
@@ -15,11 +16,12 @@ export const createDbTables = (): void => {
         db.serialize(() => {
             db.run('CREATE TABLE IF NOT EXISTS '+cameraFeedTableName+' (id INTEGER PRIMARY KEY AUTOINCREMENT, `name` TEXT, `url` TEXT, `interval` INTEGER, `wsPort` INTEGER)');
             db.run('CREATE TABLE IF NOT EXISTS '+settingsTableName+' (id INTEGER PRIMARY KEY AUTOINCREMENT, `key` TEXT, `value` TEXT, UNIQUE(`key`))');
+            db.run('CREATE TABLE IF NOT EXISTS '+recordingsTableName+' (id INTEGER PRIMARY KEY AUTOINCREMENT, `name` TEXT, `duration` INTEGER, `datetime` INTEGER)');
         });
     });
 }
 export const insertCameraFeed = (cameraFeed: CameraFeed): void => {
-    if (db === null) return;
+    if (db === null || db === undefined) return;
     db.serialize(() => {
         db.run(
             'INSERT INTO '+cameraFeedTableName+' (`name`, `url`, `interval`, `wsPort`) VALUES (?, ?, ?, ?)',
@@ -27,17 +29,8 @@ export const insertCameraFeed = (cameraFeed: CameraFeed): void => {
         );
     });
 }
-export const updateCameraFeed = (cameraFeed: CameraFeed): void => {
-    if (db === null) return;
-    db.serialize(() => {
-        db.run(
-            'UPDATE '+cameraFeedTableName+' SET `name` = ?, `url` = ?, `interval` = ?, `wsPort` = ? WHERE `id` = ?',
-            [cameraFeed.name, cameraFeed.url, cameraFeed.interval, cameraFeed.wsPort, cameraFeed.id]
-        );
-    });
-}
 export const updateCameraFeedValue = (id: number, key: string, value: string): void => {
-    if (db === null) return;
+    if (db === null || db === undefined) return;
     db.serialize(() => {
         db.run(
             'UPDATE '+cameraFeedTableName+' SET `'+key+'` = ? WHERE `id` = ?',
@@ -46,13 +39,13 @@ export const updateCameraFeedValue = (id: number, key: string, value: string): v
     });
 }
 export const deleteCameraFeed = (id: number): void => {
-    if (db === null) return;
+    if (db === null || db === undefined) return;
     db.serialize(() => {
         db.run('DELETE FROM '+cameraFeedTableName+' WHERE `id` = ?', [id]);
     });
 }
 export const getCameraFeeds = async (): Promise<CameraFeed[]> => {
-    if (db === null) return [];
+    if (db === null || db === undefined) return [];
     return await new Promise((resolve, reject) => {
         db.serialize(() => {
             db.all('SELECT * FROM ' + cameraFeedTableName, (err, rows) => {
@@ -63,8 +56,20 @@ export const getCameraFeeds = async (): Promise<CameraFeed[]> => {
     });
 }
 
+export const getCameraFeedsCount = async (): Promise<number> => {
+    if (db === null || db === undefined) return 0;
+    return await new Promise((resolve, reject) => {
+        db.serialize(() => {
+            db.get('SELECT COUNT(*) as count FROM ' + cameraFeedTableName, (err, row) => {
+                if (err) reject(err);
+                resolve(<number>row.count);
+            });
+        });
+    });
+}
+
 export const getCameraFeed = async (id: number): Promise<CameraFeed|null> => {
-    if (db === null) return null;
+    if (db === null || db === undefined) return null;
     return await new Promise((resolve, reject) => {
         db.serialize(() => {
             db.get('SELECT * FROM ' + cameraFeedTableName + ' WHERE `id` = ?', [id], (err, row) => {
@@ -80,7 +85,7 @@ export const getCameraFeed = async (id: number): Promise<CameraFeed|null> => {
 }
 
 export const getSetting = async (key: string): Promise<Setting|null> => {
-    if (db === null) return null;
+    if (db === null || db === undefined) return null;
     return await new Promise((resolve, reject) => {
         db.serialize(() => {
             db.get('SELECT value FROM ' + settingsTableName + ' WHERE `key` = ?', [key], (err, row) => {
@@ -96,14 +101,14 @@ export const getSetting = async (key: string): Promise<Setting|null> => {
 }
 
 export const setSetting = (setting: Setting): void => {
-    if (db === null) return;
+    if (db === null || db === undefined) return;
     db.serialize(() => {
         db.run('INSERT OR REPLACE INTO '+settingsTableName+' (`key`, `value`) VALUES (?, ?)', [setting.key, setting.value]);
     });
 }
 
 export const settingExists = async (key: string): Promise<boolean> => {
-    if (db === null) return false;
+    if (db === null || db === undefined) return false;
     return await new Promise((resolve, reject) => {
         db.serialize(() => {
             db.get('SELECT value FROM ' + settingsTableName + ' WHERE `key` = ?', [key], (err, row) => {
@@ -115,8 +120,37 @@ export const settingExists = async (key: string): Promise<boolean> => {
 }
 
 export const deleteSetting = (key: string): void => {
-    if (db === null) return;
+    if (db === null || db === undefined) return;
     db.serialize(() => {
         db.run('DELETE FROM '+settingsTableName+' WHERE `key` = ?', [key]);
+    });
+}
+
+export const insertRecording = (recording: Recording): void => {
+    if (db === null || db === undefined) return;
+    db.serialize(() => {
+        db.run(
+            'INSERT INTO '+recordingsTableName+' (`name`, `duration`, `datetime`) VALUES (?, ?, ?)',
+            [recording.name, recording.duration, recording.datetime]
+        );
+    });
+}
+
+export const getRecordings = async (): Promise<Recording[]> => {
+    if (db === null || db === undefined) return [];
+    return await new Promise((resolve, reject) => {
+        db.serialize(() => {
+            db.all('SELECT * FROM ' + recordingsTableName, (err, rows) => {
+                if (err) reject(err);
+                resolve(<Recording[]>rows);
+            });
+        });
+    });
+}
+
+export const deleteRecording = (id: number): void => {
+    if (db === null || db === undefined) return;
+    db.serialize(() => {
+        db.run('DELETE FROM '+recordingsTableName+' WHERE `id` = ?', [id]);
     });
 }
